+++ server/src/Ogmios/Data/Json/Shelley.hs
@@ -974,7 +974,7 @@ encodeTxIn (Ledger.TxIn txid (Ledger.TxIx ix)) = encodeObject
       , encodeTxId txid
       )
     , ( "index"
-      , encodeWord16 ix
+      , encodeWord64 ix
       )
     ]
 
@@ -1351,6 +1351,20 @@ stringifyScriptHash
     -> Text
 stringifyScriptHash (Sh.ScriptHash (CC.UnsafeHash h)) =
     encodeBase16 (fromShort h)
+
+stringifyTxId
+    :: Crypto crypto
+    => Ledger.TxId crypto
+    -> Text
+stringifyTxId (Ledger.TxId (Ledger.originalBytes -> bytes)) =
+    encodeBase16 bytes
+
+stringifyTxIn
+    :: Crypto crypto
+    => Ledger.TxIn crypto
+    -> Text
+stringifyTxIn (Ledger.TxIn txid (Ledger.TxIx ix)) =
+    stringifyTxId txid <> "#" <> show ix
 
 stringifyVKey
     :: Crypto crypto